fix for aftereach

+++ Tests/Admin/RsEncryptionKey.Tests.ps1
@@ -38,8 +38,6 @@ Describe "RsEncryptionKey" {
 
             Backup-RsEncryptionKey -Password $keyPassword -KeyPath $keyPath -Confirm:$false -Verbose -ReportServerInstance PBIRS -ReportServerVersion SQLServer2017
             Restore-RsEncryptionKey -Password $keyPassword -KeyPath $keyPath -Confirm:$false -Verbose -ReportServerInstance PBIRS -ReportServerVersion SQLServer2017
-            Stop-Service PowerBIReportServer
-            Start-Service PowerBIReportServer
 
             Test-AccessToEncryptedContent -ExpectedDataSource $dataSource
         }
@@ -55,9 +53,6 @@ Describe "RsEncryptionKey" {
             Backup-RsEncryptionKey -Password $keyPassword -KeyPath $keyPath -Confirm:$false -Verbose -ReportServerInstance PBIRS -ReportServerVersion SQLServer2017
             Restore-RsEncryptionKey -Password $keyPassword -KeyPath $keyPath -Confirm:$false -Verbose -ReportServerInstance PBIRS -ReportServerVersion SQLServer2017
 
-            Stop-Service PowerBIReportServer
-            Start-Service PowerBIReportServer
-
             Test-AccessToEncryptedContent -ExpectedDataSource $dataSource
         }
 
@@ -66,7 +61,9 @@ Describe "RsEncryptionKey" {
         }
 
         AfterEach {
-            Remove-RsCatalogItem -Path $itemsToClean
+            if($itemsToClean.Count -gt 0){
+                Remove-RsCatalogItem -Path $itemsToClean
+            }
         }
     }
 }
